Moved to database
Created a new package called "database" and removed the "mongo"  package.  Also renamed the "add" function to "create" and the "query" function to "read".
