+++ database/ParseCSV.go
@@ -1,4 +1,4 @@
-package mongo
+package database
 
 import (
 	"bufio"
@@ -11,11 +11,11 @@ import (
 )
 
 //ParseCSV is used to imort a CSV file.
-func ParseCSV(mf multipart.File) []Device {
+func ParseCSV(mf multipart.File) []MongoDevice {
 	fmt.Println("in parse CSV")
 	csvFile := mf
 	reader := csv.NewReader(bufio.NewReader(csvFile))
-	var devices []Device
+	var devices []MongoDevice
 	for {
 		line, error := reader.Read()
 		if error == io.EOF {
@@ -23,7 +23,7 @@ func ParseCSV(mf multipart.File) []Device {
 		} else if error != nil {
 			log.Fatal(error)
 		}
-		devices = append(devices, Device{
+		devices = append(devices, MongoDevice{
 			Hostname:   line[0],
 			IPAddress:  line[1],
 			DeviceType: line[2],
