+++ page/QueryPostHandler.go
@@ -1,10 +1,9 @@
 package page
 
 import (
+	"go_mongo/database"
 	"log"
 	"net/http"
-
-	"go_mongo/mongo"
 )
 
 //QueryPostHandler calls the query.gohtml after recieving a post. URL is localhost/query.
@@ -15,13 +14,13 @@ func QueryPostHandler(res http.ResponseWriter, req *http.Request) {
 		log.Fatalln(err)
 	}
 	andOr := req.FormValue("anyOr")
-	queryDevice := mongo.Device{
+	queryDevice := database.MongoDevice{
 		Hostname:   req.FormValue("hostname"),
 		IPAddress:  req.FormValue("ipAddress"),
 		DeviceType: req.FormValue("deviceType"),
 	}
 	//
-	deviceList := mongo.ListDevice(&queryDevice, &andOr, res, req)
+	deviceList := database.Read(&queryDevice, &andOr, res, req)
 	err = tpl.ExecuteTemplate(res, "query.gohtml", deviceList)
 	if err != nil {
 		log.Fatalln(err)
