+++ page/QueryResponseHandler.go
@@ -1,10 +1,9 @@
 package page
 
 import (
+	"go_mongo/database"
 	"log"
 	"net/http"
-
-	"go_mongo/mongo"
 )
 
 //QueryResponseHandler is used for query.
@@ -14,13 +13,13 @@ func QueryResponseHandler(res http.ResponseWriter, req *http.Request) {
 		log.Fatalln(err)
 	}
 	andOr := req.FormValue("anyOr")
-	queryDevice := mongo.Device{
+	queryDevice := database.MongoDevice{
 		Hostname:   req.FormValue("hostname"),
 		IPAddress:  req.FormValue("ipAddress"),
 		DeviceType: req.FormValue("deviceType"),
 	}
 	//
-	deviceList := mongo.ListDevice(&queryDevice, &andOr, res, req)
+	deviceList := database.Read(&queryDevice, &andOr, res, req)
 	err = tpl.ExecuteTemplate(res, "queryResponse.gohtml", deviceList)
 	if err != nil {
 		log.Fatalln(err)
